+++ cs-lib/lubx/Text.cs
@@ -22,7 +22,7 @@ public class TextGlyph
     public int h;
     public int xoff;
     public int yoff;
-    public double advance;
+    public float advance;
 }
 
 /// <summary>固定サイズの動的 glyph atlas + 1行テキスト描画。Font.font_glyph を
@@ -34,16 +34,16 @@ public class TextGlyph
 public class Text
 {
     /// <summary>1em のピクセル数 (コンストラクタ指定)。</summary>
-    public double px;
+    public float px;
 
     /// <summary>ベースラインから上端まで (px、正)。読み取り専用。</summary>
-    public double ascent;
+    public float ascent;
 
     /// <summary>ベースラインから下端まで (px、負)。読み取り専用。</summary>
-    public double descent;
+    public float descent;
 
     /// <summary>行送り (px)。読み取り専用。</summary>
-    public double lineHeight;
+    public float lineHeight;
 
     private string ttf;
     private Atlas atlas;
@@ -58,7 +58,7 @@ public class Text
     private int rowH = 0;
 
     /// <summary>atlasSize 省略で 256。</summary>
-    public Text(string key, string ttf, double px, int? atlasSize = null)
+    public Text(string key, string ttf, float px, int? atlasSize = null)
     {
         this.ttf = ttf;
         this.px = px;
@@ -155,9 +155,9 @@ public class Text
     }
 
     /// <summary>1行の描画幅 (px)。scale 省略で 1.0。</summary>
-    public double width(string s, double? scale = null)
+    public float width(string s, float? scale = null)
     {
-        var sum = 0.0;
+        var sum = 0.0f;
         var prev = -1;
         eachCodepoint(s, cp =>
         {
@@ -169,15 +169,15 @@ public class Text
             sum += g.advance;
             prev = cp;
         });
-        return sum * (scale ?? 1.0);
+        return sum * (scale ?? 1.0f);
     }
 
     /// <summary>1行描く。(x, y) はベースライン左端 (論理 px、左上原点)。
     /// scale 省略で 1.0。</summary>
-    public void draw(SpriteBatch batch, string s, double x, double y,
-        Color? tint = null, double? scale = null)
+    public void draw(SpriteBatch batch, string s, float x, float y,
+        Color? tint = null, float? scale = null)
     {
-        double sc = scale ?? 1.0;
+        float sc = scale ?? 1.0f;
         var pen = x;
         var prev = -1;
         eachCodepoint(s, cp =>
